# services/UserManagement/user-manager-server/src/main/resources/api-spec.yaml
swagger: "2.0"
info:
  description: "This is the user manager API for AMT Project Two"
  version: "1.0.0"
  title: "Users"
host: localhost:8080
basePath: /api
schemes:
  - "http"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /authenticate:
    post:
      tags:
        - "auth"
      summary: "Authentication"
      description: ""
      operationId: createAuthenticationToken
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: '#/definitions/JwtRequest'
      responses:
        201:
          description: token generated
          schema:
            $ref: '#/definitions/JwtResponse'
        403:
          description: blocked
          schema:
            type: object
  /users:
    post:
      security:
      - Bearer: []
      tags:
        - "users"
      summary: "Create a new user"
      description: ""
      operationId: createUser
      consumes:
        - application/json
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserInput'
      responses:
        201:
          description: created
          schema:
            type: object
        403:
          description: blocked
          schema:
            type: object
        400:
          description: wrong format
          schema:
            type: object
        409:
          description: duplicate ressource
          schema:
            type: object
    get:
      security:
      - Bearer: []
      tags:
        - "users"
      summary: "Get the list of all users"
      description: ""
      operationId: "getUsers"
      produces:
        - "application/json"
      responses:
        '401':
          description: "Unauthorized"
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: '#/definitions/UserOutput'
        400:
          description: wrong format
          schema:
            type: object
        403:
          description: blocked
          schema:
            type: object
        404:
          description: ressource not found
          schema:
            type: object
  /users/{email}:
    get:
      security:
      - Bearer: []
      tags:
        - "users"
      summary: "Gets a user by his email"
      description: ""
      operationId: "getUserById"
      produces:
        - "application/json"
      parameters:
        - in: path
          type: string
          required: true
          name: email
          description: The email we need to Id the user
      responses:
        200:
          description: "successful operation"
          schema:
            type: object
        400:
          description: wrong format
          schema:
            type: object
        403:
          description: blocked
          schema:
            type: object
        404:
          description: ressource not found
          schema:
            type: object
    patch:
      security:
      - Bearer: []
      tags:
        - "users"
      summary: "Updates the ressource"
      description: ""
      operationId: "updateUser"
      produces:
        - 'application/json'
      consumes:
        - 'application/json'
      parameters:
        - name: password
          in: header
          required: true
          type: string
        - name: email
          in: path
          type: string
          required: true
          description: The email we need to Id the user
      responses:
        '201':
          description: updated
          schema:
            type: object
        400:
          description: wrong format
          schema:
            type: object
        403:
          description: blocked
          schema:
            type: object
        404:
          description: ressource not found
          schema:
            type: object
    delete:
      security:
      - Bearer: []
      tags:
        - "users"
      summary: "Deletes a given ressource"
      operationId: "deleteUser"
      parameters:
        - name: email
          in: path
          required: true
          type: string
          description: The email we need to Id the user
      responses:
        201:
          description: deleted
          schema:
            type: object
        400:
          description: wrong format
          schema:
            type: object
        403:
          description: blocked
          schema:
            type: object
        404:
          description: ressource not found
          schema:
            type: object
definitions:
  User:
    type: "object"
    properties:
      email:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      isAdmin:
        type: "integer"
      isBlocked:
        type: "integer"
  UserOutput:
    allOf:
      - type: "object"
      - $ref: "#/definitions/User"
  UserInput:
    allOf:
      - $ref: "#/definitions/User"
      - type: "object"
      - properties:
          password:
            type: "string"
  JwtRequest:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

  JwtResponse:
      type: "object"
      properties:
        token:
          type: "string"
#components:
#  responses:
#    UnauthorizedError:
#      description: Access token is missing or invalid